requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.4.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pytest==8.3.4
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9